+++ docs/root/intro/version_history.rst
@@ -3,7 +3,6 @@ Version history
 
 1.7.0 (Pending)
 ===============
-
 * access log: added ability to log response trailers.
 * access log: added ability to format START_TIME.
 * access log: added DYNAMIC_METADATA :ref:`access log formatter <config_access_log_format>`.
@@ -36,6 +35,8 @@ Version history
 * debug: added symbolized stack traces (where supported).
 * grpc: support added for the full set of :ref:`Google gRPC call credentials
   <envoy_api_msg_core.GrpcService.GoogleGrpc.CallCredentials>`.
+* gzip filter: added :ref:`stats <gzip-statistics>` to the filter.
+* gzip filter: sending *accept-encoding* header as *identity* no longer compresses the payload.
 * health check: added ability to set :ref:`additional HTTP headers
   <envoy_api_field_core.HealthCheck.HttpHealthCheck.request_headers_to_add>` for HTTP health check.
 * health check: added support for EDS delivered :ref:`endpoint health status